// part9/patientor-backend/src/utils.ts
import z from 'zod';

import { Gender, HealthCheckRating } from './types';

export const NewPatientSchema = z.object({
  name: z.string(),
  dateOfBirth: z.iso.date(),
  ssn: z.string().regex(/^\d{6}-\d{2,3}[A-Z0-9]$/),
  gender: z.enum(Gender),
  occupation: z.string(),
  entries: z.any(),
});

const NewBaseEntrySchema = z.object({
  description: z.string().min(10),
  date: z.iso.date(),
  specialist: z.string().min(5),
  diagnosisCodes: z.array(z.string()).optional(),
});

const HealthCheckRatingSchema = z.enum(HealthCheckRating);

const HealthCheckEntrySchema = NewBaseEntrySchema.extend({
  type: z.literal('HealthCheck'),
  healthCheckRating: HealthCheckRatingSchema,
});

const OccupationalHealthcareEntrySchema = NewBaseEntrySchema.extend({
  type: z.literal('OccupationalHealthcare'),
  employerName: z.string(),
  sickLeave: z
    .object({
      startDate: z.string(),
      endDate: z.string(),
    })
    .optional(),
});

const HospitalEntrySchema = NewBaseEntrySchema.extend({
  type: z.literal('Hospital'),
  discharge: z.object({
    date: z.string(),
    criteria: z.string(),
  }),
});

export const NewEntrySchema = z.discriminatedUnion('type', [
  HealthCheckEntrySchema,
  OccupationalHealthcareEntrySchema,
  HospitalEntrySchema,
]);
